+++ R/edgelists.R
@@ -19,13 +19,13 @@ get_edgelist <- function(dat, network) {
     if (!network %in% seq_len(dat$num.nw)) {
       stop("There is no network '", network, "' to get an edgelist from")
     }
-    el <- dat[["el"]][[network]]
+    el <- dat$run$el[[network]]
   } else {
     if (!network %in% seq_len(dat$num.nw)) {
       stop("There is no network '", network, "' to get an edgelist from")
     }
     at <- get_current_timestep(dat)
-    el <- networkDynamic::get.dyads.active(dat[["nw"]][[network]], at = at)
+    el <- networkDynamic::get.dyads.active(dat$run$nw[[network]], at = at)
   }
 
   return(el)
@@ -180,20 +180,10 @@ get_cumulative_edgelist <- function(dat, network) {
          `cumulative.edgelist` control setting is set to `FALSE`.")
   }
 
-  if (length(dat$el.cuml) >= network) {
-    el_cuml <- dat[["el.cuml"]][[network]]
-  } else {
-    el_cuml <- NULL
-  }
-
-  if (is.null(el_cuml)) {
-    el_cuml <- tibble::tibble(
-      head  = numeric(0),
-      tail  = numeric(0),
-      start = numeric(0),
-      stop  = numeric(0)
-    )
-  }
+  el_cuml <- dplyr::bind_rows(
+    get_raw_elcuml(dat, network, active = FALSE),
+    get_raw_elcuml(dat, network, active = TRUE)
+  )
 
   return(el_cuml)
 }
@@ -224,7 +214,16 @@ update_cumulative_edgelist <- function(dat, network, truncate = 0) {
   }
 
   el <- get_edgelist(dat, network)
-  el_cuml <- get_cumulative_edgelist(dat, network)
+  el_cuml_cur <- get_raw_elcuml(dat, network, active = TRUE)
+  el_cuml_hist <- get_raw_elcuml(dat, network, active = FALSE)
+
+  at <- get_current_timestep(dat)
+
+  # truncate el_cuml_hist
+  if (truncate != Inf && truncate != 0) {
+    rel.age <- at - el_cuml_hist$stop
+    el_cuml_hist <- el_cuml_hist[rel.age <= truncate, ]
+  }
 
   el <- tibble::tibble(
     head = get_unique_ids(dat, el[, 1]),
@@ -232,27 +231,32 @@ update_cumulative_edgelist <- function(dat, network, truncate = 0) {
     current = TRUE
   )
 
-  el_cuml <- dplyr::full_join(el_cuml, el, by = c("head", "tail"))
+  el_cuml_cur <- dplyr::full_join(el_cuml_cur, el, by = c("head", "tail"))
 
-  at <- get_current_timestep(dat)
-
-  new_edges <- is.na(el_cuml[["start"]])
-  if (any(new_edges)) {
-    el_cuml[new_edges, ][["start"]] <- at
+  # new edges
+  new_edges_ids <- which(is.na(el_cuml_cur$start))
+  if (length(new_edges_ids) > 0) {
+    el_cuml_cur[new_edges_ids, ]$start <- at
   }
 
-  terminated_edges <- is.na(el_cuml[["current"]]) & is.na(el_cuml[["stop"]])
-  if (any(terminated_edges)) {
-    el_cuml[terminated_edges, ][["stop"]] <- at - 1
+  # terminated edges
+  terminated_edges_ids <- which(is.na(el_cuml_cur$current))
+  el_cuml_cur$current <- NULL
+
+  if (length(terminated_edges_ids) > 0) {
+    el_cuml_term <- el_cuml_cur[terminated_edges_ids, ]
+
+    # with truncate == 0, don't save any historic edges
+    if (truncate != 0) {
+      el_cuml_term$stop <- at - 1
+      el_cuml_hist <- dplyr::bind_rows(el_cuml_hist, el_cuml_term)
+    }
+
+    el_cuml_cur <- el_cuml_cur[-terminated_edges_ids, ]
   }
 
-  if (truncate != Inf) {
-    rel.age <- at - el_cuml[["stop"]]
-    rel.age <- ifelse(is.na(rel.age), 0, rel.age)
-    el_cuml <- el_cuml[rel.age <= truncate, ]
-  }
-
-  dat[["el.cuml"]][[network]] <- el_cuml[, c("head", "tail", "start", "stop")]
+  dat <- set_raw_elcuml(dat, network, el_cuml_cur, active = TRUE)
+  dat <- set_raw_elcuml(dat, network, el_cuml_hist, active = FALSE)
 
   return(dat)
 }
@@ -383,3 +387,42 @@ get_cumulative_degree <- function(dat, index_posit_ids, networks = NULL,
     dplyr::mutate(index = get_posit_ids(dat, .data$index)) |>
     dplyr::select(index_pid = "index", "degree")
 }
+
+# Helper functions to get and set the cumulative edgelists
+#
+# Cumulative edgelists are split in to:
+#   - el_cuml_cur: for the edges not yet disolved
+#   - el_cuml_hist: for the edges where the start and stop time are known
+#   (dissolved edges)
+get_raw_elcuml <- function(dat, network, active) {
+  loc <- if (active) "el_cuml_cur" else "el_cuml_hist"
+
+  if (length(dat$run[[loc]]) >= network) {
+    el_cuml <- dat$run[[loc]][[network]]
+  } else {
+    el_cuml <- NULL
+  }
+
+  if (is.null(el_cuml)) {
+    el_cuml <- empty_el_cuml()
+  }
+
+  return(el_cuml)
+}
+
+set_raw_elcuml <- function(dat, network, el_cuml, active) {
+  loc <- if (active) "el_cuml_cur" else "el_cuml_hist"
+  dat$run[[loc]][[network]] <- el_cuml
+  return(dat)
+}
+
+# template for the cumulative edgelists
+empty_el_cuml <- function() {
+  tibble::tibble(
+    head  = numeric(0),
+    tail  = numeric(0),
+    start = numeric(0),
+    stop  = numeric(0)
+  )
+}
+