+++ R/net.inputs.R
@@ -714,6 +714,9 @@ init.net <- function(i.num, r.num, i.num.g2, r.num.g2,
 #'        specification.
 #' @param save.network If `TRUE`, networkDynamic or networkLite object is saved at simulation end.
 #' @param save.transmat If `TRUE`, complete transmission matrix is saved at simulation end.
+#' @param save.run If `TRUE`, the `run` sublist of `dat` is saved, allowing a
+#'   simulation to restart from this output.
+#' @param save.cumulative.edgelist If `TRUE`, the `cumulative.edgelist` is saved at simulation end.
 #' @param save.other A character vector of elements on the `netsim_dat` main data list to save out
 #'        after each simulation. One example for base models is the attribute list, `"attr"`, at
 #'        the final time step.
@@ -867,6 +870,8 @@ control.net <- function(type,
                         nwstats.formula = "formation",
                         save.transmat = TRUE,
                         save.network,
+                        save.run = FALSE,
+                        save.cumulative.edgelist = FALSE,
                         save.other,
                         verbose = TRUE,
                         verbose.int = 1,
@@ -1144,13 +1149,9 @@ crosscheck.net <- function(x, param, init, control) {
           stop("x must be a netsim object if control setting start > 1",
                call. = FALSE)
         }
-        if (is.null(x[["attr"]])) {
-          stop("x must contain attr to restart simulation, see save.other ",
+        if (is.null(x[["run"]])) {
+          stop("x must contain `run` to restart simulation, see `save.run` ",
                "control setting", call. = FALSE)
-        }
-        if (is.null(x[["network"]]) && control[["tergmLite"]] == FALSE) {
-          stop("x must contain network object to restart simulation when ",
-               "tergmLite == FALSE, ", call. = FALSE)
         }
         if (control[["nsteps"]] < control[["start"]]) {
           stop("control setting nsteps must be > control setting start in ",