+++ R/net.mod.prevalence.R
@@ -26,24 +26,24 @@ prevalence.net <- function(dat, at) {
   groups <- get_param(dat, "groups")
 
   # Subset attr to active == 1
-  l <- lapply(seq_along(dat$attr), function(x) dat$attr[[x]][active == 1])
-  names(l) <- names(dat$attr)
+  attr_list <- get_attr_list(dat)
+  l <- lapply(seq_along(attr_list), function(x) attr_list[[x]][active == 1])
+  names(l) <- names(attr_list)
   l$active <- l$infTime <- NULL
 
   status <- l$status
 
   ## Subsetting for epi.by control
-  eb <- !is.null(dat$control$epi.by)
-  if (eb == TRUE) {
-    ebn <- get_control(dat, "epi.by")
-    ebv <- dat$temp$epi.by.vals
+  ebn <- get_control(dat, "epi.by", override.null.error = TRUE)
+  if (!is.null(ebn)) {
+    ebv <- dat$run$epi.by.vals
     ebun <- paste0(".", ebn, ebv)
     assign(ebn, l[[ebn]])
   }
 
   if (groups == 1) {
     dat <- set_epi(dat, "s.num", at, sum(status == "s"))
-    if (eb == TRUE) {
+    if (!is.null(ebn)) {
       for (i in seq_along(ebun)) {
         ebn.temp <- paste0("s.num", ebun[i])
         dat <- set_epi(dat, ebn.temp, at, sum(status == "s" &
@@ -52,7 +52,7 @@ prevalence.net <- function(dat, at) {
     }
 
     dat <- set_epi(dat, "i.num", at, sum(status == "i"))
-    if (eb == TRUE) {
+    if (!is.null(ebn)) {
       for (i in seq_along(ebun)) {
         ebn.temp <- paste0("i.num", ebun[i])
         dat <- set_epi(dat, ebn.temp, at, sum(status == "i" &
@@ -62,7 +62,7 @@ prevalence.net <- function(dat, at) {
 
     if (type == "SIR") {
       dat <- set_epi(dat, "r.num", at, sum(status == "r"))
-      if (eb == TRUE) {
+      if (!is.null(ebn)) {
         for (i in seq_along(ebun)) {
           ebn.temp <- paste0("r.num", ebun[i])
           dat <- set_epi(dat, ebn.temp, at, sum(status == "r" &
@@ -71,7 +71,7 @@ prevalence.net <- function(dat, at) {
       }
     }
     dat <- set_epi(dat, "num", at, length(status))
-    if (eb == TRUE) {
+    if (!is.null(ebn)) {
       for (i in seq_along(ebun)) {
         ebn.temp <- paste0("num", ebun[i])
         dat <- set_epi(dat, ebn.temp, at, sum(get(ebn) == ebv[i]))
@@ -84,7 +84,7 @@ prevalence.net <- function(dat, at) {
     group <- group[active == 1]
 
     dat <- set_epi(dat, "s.num", at, sum(status == "s" & group == 1))
-    if (eb == TRUE) {
+    if (!is.null(ebn)) {
       for (i in seq_along(ebun)) {
         ebn.temp <- paste0("s.num", ebun[i])
         dat <- set_epi(dat, ebn.temp, at, sum(status == "s" &
@@ -93,7 +93,7 @@ prevalence.net <- function(dat, at) {
       }
     }
     dat <- set_epi(dat, "i.num", at, sum(status == "i" & group == 1))
-    if (eb == TRUE) {
+    if (!is.null(ebn)) {
       for (i in seq_along(ebun)) {
         ebn.temp <- paste0("i.num", ebun[i])
         dat <- set_epi(dat, ebn.temp, at, sum(status == "i" &
@@ -103,7 +103,7 @@ prevalence.net <- function(dat, at) {
     }
     if (type == "SIR") {
       dat <- set_epi(dat, "r.num", at, sum(status == "r" & group == 1))
-      if (eb == TRUE) {
+      if (!is.null(ebn)) {
         for (i in seq_along(ebun)) {
           ebn.temp <- paste0("r.num", ebun[i])
           dat <- set_epi(dat, ebn.temp, at, sum(status == "r" &
@@ -113,8 +113,7 @@ prevalence.net <- function(dat, at) {
       }
     }
     dat <- set_epi(dat, "num", at, sum(group == 1))
-    dat$epi$num[at] <- sum(group == 1)
-    if (eb == TRUE) {
+    if (!is.null(ebn)) {
       for (i in seq_along(ebun)) {
         ebn.temp <- paste0("num", ebun[i])
         dat <- set_epi(dat, ebn.temp, at, sum(group == 1 &
@@ -122,7 +121,7 @@ prevalence.net <- function(dat, at) {
       }
     }
     dat <- set_epi(dat, "s.num.g2", at, sum(status == "s" & group == 2))
-    if (eb == TRUE) {
+    if (!is.null(ebn)) {
       for (i in seq_along(ebun)) {
         ebn.temp <- paste0("s.num.g2", ebun[i])
         dat <- set_epi(dat, ebn.temp, at, sum(status == "s" &
@@ -131,7 +130,7 @@ prevalence.net <- function(dat, at) {
       }
     }
     dat <- set_epi(dat, "i.num.g2", at, sum(status == "i" & group == 2))
-    if (eb == TRUE) {
+    if (!is.null(ebn)) {
       for (i in seq_along(ebun)) {
         ebn.temp <- paste0("i.num.g2", ebun[i])
         dat <- set_epi(dat, ebn.temp, at, sum(status == "i" &
@@ -141,7 +140,7 @@ prevalence.net <- function(dat, at) {
     }
     if (type == "SIR") {
       dat <- set_epi(dat, "r.num.g2", at, sum(status == "r" & group == 2))
-      if (eb == TRUE) {
+      if (!is.null(ebn)) {
         for (i in seq_along(ebun)) {
           ebn.temp <- paste0("r.num.g2", ebun[i])
           dat <- set_epi(dat, ebn.temp, at, sum(status == "r" &
@@ -151,7 +150,7 @@ prevalence.net <- function(dat, at) {
       }
     }
     dat <- set_epi(dat, "num.g2", at, sum(group == 2))
-    if (eb == TRUE) {
+    if (!is.null(ebn)) {
       for (i in seq_along(ebun)) {
         ebn.temp <- paste0("num.g2", ebun[i])
         dat <- set_epi(dat, ebn.temp, at, sum(group == 2 &