+++ R/update.R
@@ -180,7 +180,7 @@ depart_nodes <- function(dat, departures) {
   if (length(departures) > 0) {
     if (get_control(dat, "tergmLite") == FALSE) {
       for (network in seq_len(dat$num.nw)) {
-        dat$nw[[network]] <- deactivate.vertices(dat$nw[[network]],
+        dat$run$nw[[network]] <- deactivate.vertices(dat$run$nw[[network]],
                                                  onset = get_current_timestep(dat),
                                                  terminus = Inf,
                                                  v = departures,
@@ -189,12 +189,12 @@ depart_nodes <- function(dat, departures) {
     } else {
       dat <- delete_attr(dat, departures)
       for (network in seq_len(dat$num.nw)) {
-        dat$el[[network]] <- delete_vertices(dat$el[[network]], departures)
-        dat$net_attr[[network]][["n"]] <- dat$net_attr[[network]][["n"]] - length(departures)
+        dat$run$el[[network]] <- delete_vertices(dat$run$el[[network]], departures)
+        dat$run$net_attr[[network]][["n"]] <- dat$run$net_attr[[network]][["n"]] - length(departures)
 
         if (get_network_control(dat, network, "tergmLite.track.duration") == TRUE) {
-          dat$net_attr[[network]][["lasttoggle"]] <-
-            delete_vertices(dat$net_attr[[network]][["lasttoggle"]], departures)
+          dat$run$net_attr[[network]][["lasttoggle"]] <-
+            delete_vertices(dat$run$net_attr[[network]][["lasttoggle"]], departures)
         }
       }
     }
@@ -227,15 +227,15 @@ arrive_nodes <- function(dat, nArrivals) {
   if (nArrivals > 0) {
     if (get_control(dat, "tergmLite") == FALSE) {
       for (network in seq_len(dat$num.nw)) {
-        dat$nw[[network]] <- add.vertices.active(dat$nw[[network]],
+        dat$run$nw[[network]] <- add.vertices.active(dat$run$nw[[network]],
                                                  nv = nArrivals,
                                                  onset = get_current_timestep(dat),
                                                  terminus = Inf)
       }
     } else {
       for (network in seq_len(dat$num.nw)) {
-        dat$el[[network]] <- add_vertices(dat$el[[network]], nv = nArrivals)
-        dat$net_attr[[network]][["n"]] <- dat$net_attr[[network]][["n"]] + nArrivals
+        dat$run$el[[network]] <- add_vertices(dat$run$el[[network]], nv = nArrivals)
+        dat$run$net_attr[[network]][["n"]] <- dat$run$net_attr[[network]][["n"]] + nArrivals
       }
     }
   }